+++ test/Sentry.Tests/BindableSentryOptionsTests.cs
@@ -5,12 +5,6 @@ namespace Sentry.Tests;
 
 public class BindableSentryOptionsTests : BindableTests<SentryOptions>
 {
-    public BindableSentryOptionsTests() : base(
-        nameof(SentryOptions.ExperimentalMetrics)
-    )
-    {
-    }
-
     [Fact]
     public void BindableProperties_MatchOptionsProperties()
     {